+++ README.md
@@ -1,2 +1,4 @@
 # machete
 Open source tool for GMs to track, run, plan, and do analytics on their campaign.
+
+You can run it with `cargo run` or `cargo watch -x run`
