+++ src/lib.rs
@@ -4,6 +4,7 @@ pub use app::MainApp;
 pub mod apps;
 pub mod models;
 pub mod settings_panel;
+pub mod ui_models;
 pub mod update_context;
 pub mod utils;
 pub mod widgets;
